requirements.txt
requests
telegram
python-telegram-bot
Pillow